+++ packages/remote-react-renderer/src/RemoteRendererBrowser.tsx
@@ -11,12 +11,19 @@ import {
   type NavigationState,
   type RemoteExtBridgeConnectionApi,
 } from "@mittwald/flow-remote-core";
-import { usePromise } from "@mittwald/react-use-promise";
+import { refresh, usePromise } from "@mittwald/react-use-promise";
 import {
   RemoteReceiver,
   RemoteRootRenderer,
 } from "@mittwald/remote-dom-react/host";
-import { type CSSProperties, type FC, useMemo, useRef, useState } from "react";
+import {
+  type CSSProperties,
+  type FC,
+  useEffect,
+  useMemo,
+  useRef,
+  useState,
+} from "react";
 import { useLanguage } from "@mittwald/flow-react-components";
 import type { HostConfig } from "@mittwald/flow-core";
 
@@ -143,6 +150,18 @@ export const RemoteRendererBrowser: FC<RemoteRendererBrowserProps> = (
 
   const awaitLoadingPromise = connectionSrc === src;
 
+  const loaderResourceTag = `mittwald/remote-react-renderer/connect/${src}`;
+
+  // refresh on remount
+  useEffect(
+    () => () => {
+      refresh({
+        tag: loaderResourceTag,
+      });
+    },
+    [src],
+  );
+
   usePromise(
     async () => {
       await overallLoading();
@@ -151,6 +170,7 @@ export const RemoteRendererBrowser: FC<RemoteRendererBrowserProps> = (
     awaitLoadingPromise ? [] : null,
     {
       loaderId: src,
+      tags: [loaderResourceTag],
     },
   );